moved tests to a different class and fixed bad test
