+++ code/truth/TruthAssignment.java
@@ -1,6 +1,7 @@
 package truth;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
 
 import java.util.HashMap;
 
@@ -78,36 +79,5 @@ public class TruthAssignment {
     @Override
     public String toString() {
         return data.toString();
-    }
-
-    @Test
-    public void testTruthAssignment() {
-        assertDoesNotThrow(() -> new TruthAssignment(new String[]{"A", "a", "c", "B"}, new boolean[]{true, false, false, true}));
-        assertDoesNotThrow(() -> new TruthAssignment(new String[]{}, new boolean[]{}));
-        assertDoesNotThrow(() -> new TruthAssignment(new String[]{"aa", "ba", "ac", "d", "AA"}, new boolean[]{false, false, false, false, false}));
-
-        assertThrows(IllegalArgumentException.class, () -> new TruthAssignment(new String[]{"aa", "aa"}, new boolean[]{true, false}));
-        assertThrows(IllegalArgumentException.class, () -> new TruthAssignment(new String[]{"a", "A", "a"}, new boolean[]{true, false, true}));
-
-        assertThrows(IllegalArgumentException.class, () -> new TruthAssignment(new String[]{"a", "A"}, new boolean[]{true}));
-        assertThrows(IllegalArgumentException.class, () -> new TruthAssignment(new String[]{}, new boolean[]{true}));
-    }
-
-    @Test
-    public void testMethods() {
-        TruthAssignment ta = new TruthAssignment(new String[]{"a", "b", "c"}, new boolean[]{true, false, true});
-
-        assertTrue(ta.getValue("a"));
-        assertFalse(ta.getValue("b"));
-        assertThrows(IllegalArgumentException.class, () -> ta.getValue("d"));
-
-        assertDoesNotThrow(() -> ta.setValue("a", false));
-        assertFalse(ta.getValue("a"));
-        assertThrows(IllegalArgumentException.class, () -> ta.setValue("d", true));
-
-        assertDoesNotThrow(() -> ta.addValue("f", true));
-        assertTrue(ta.getValue("f"));
-        assertThrows(IllegalArgumentException.class, () -> ta.setValue("a", true));
-        assertFalse(ta.getValue("a"));
     }
 }
